+++ functions/handleCommand.js
@@ -2,7 +2,10 @@ module.exports = async (client, msg, command, args = undefined) => {
   const validCommand = this.getCommand(client, command);
   if (!validCommand) return;
   const response = this.runInhibitors(client, msg, validCommand);
-  if (response) return msg.reply(response);
+  if (response) {
+    if (typeof response === "string") return msg.reply(response);
+    return;
+  }
   try {
     const params = await client.funcs.usage.run(client, msg, validCommand, args);
     validCommand.run(client, msg, params);